fix: hide typed response accumulation internals
