+++ scripts/utils/postprocess-files.cjs
@@ -46,6 +46,9 @@ async function postprocess() {
     }
   }
 
+  const packageJson = /** @type {{ exports: Record<string, unknown> }} */ (
+    JSON.parse(await fs.promises.readFile('dist/package.json', 'utf-8'))
+  );
   const newExports = {
     '.': {
       require: {
@@ -55,6 +58,7 @@ async function postprocess() {
       types: './index.d.mts',
       default: './index.mjs',
     },
+    ...Object.fromEntries(Object.entries(packageJson.exports).filter(([, value]) => value === null)),
   };
 
   for (const entry of await fs.promises.readdir(distDir, { withFileTypes: true })) {
@@ -89,14 +93,9 @@ async function postprocess() {
   await fs.promises.writeFile(
     'dist/package.json',
     JSON.stringify(
-      Object.assign(
-        /** @type {Record<String, unknown>} */ (
-          JSON.parse(await fs.promises.readFile('dist/package.json', 'utf-8'))
-        ),
-        {
-          exports: newExports,
-        },
-      ),
+      Object.assign(packageJson, {
+        exports: newExports,
+      }),
       null,
       2,
     ),
