+++ src/lib/responses/ResponseStream.ts
@@ -13,21 +13,10 @@ import type OpenAI from '../../index';
 import { EventStream } from '../EventStream';
 import type { BaseEvents } from '../EventStream';
 import type { ResponseFunctionCallArgumentsDeltaEvent, ResponseTextDeltaEvent } from './EventTypes';
-import { accumulateResponse } from './ResponseAccumulator';
+import { accumulateResponseWithContext, createResponseContext } from './_ResponseAccumulator';
 import type { ParseableToolsParams } from '../ResponsesParser';
 import { maybeParseResponse } from '../ResponsesParser';
 import { Stream } from '../../streaming';
-
-type ResponseStreamAccumulatorContext = {
-  canonicalSnapshot: Response | undefined;
-  outputTextLengths: WeakMap<Response['output'][number], number>;
-};
-
-const accumulateStreamResponse = accumulateResponse as (
-  event: ResponseStreamEvent,
-  snapshot: Response | undefined,
-  context: ResponseStreamAccumulatorContext,
-) => Response;
 
 /** Parameters for starting a new response stream or replaying an existing response. */
 export type ResponseStreamParams = ResponseCreateAndStreamParams | ResponseStreamByIdParams;
@@ -104,10 +93,7 @@ export class ResponseStream<ParsedT = null>
   #params: ResponseStreamingParams | null;
   #currentResponseSnapshot: Response | undefined;
   #finalResponse: ParsedResponse<ParsedT> | undefined;
-  #accumulatorContext: ResponseStreamAccumulatorContext = {
-    canonicalSnapshot: undefined,
-    outputTextLengths: new WeakMap(),
-  };
+  #accumulatorContext = createResponseContext();
 
   /** Creates an unstarted stream, retaining request parameters for structured-output parsing. */
   constructor(params: ResponseStreamingParams | null) {
@@ -143,7 +129,7 @@ export class ResponseStream<ParsedT = null>
       return;
     }
     this.#currentResponseSnapshot = undefined;
-    this.#accumulatorContext = { canonicalSnapshot: undefined, outputTextLengths: new WeakMap() };
+    this.#accumulatorContext = createResponseContext();
   }
 
   #addEvent(this: ResponseStream<ParsedT>, event: ResponseStreamEvent, starting_after: number | null) {
@@ -165,7 +151,11 @@ export class ResponseStream<ParsedT = null>
       throw new APIError(undefined, error, event.message, undefined);
     }
 
-    const response = accumulateStreamResponse(event, this.#currentResponseSnapshot, this.#accumulatorContext);
+    const response = accumulateResponseWithContext(
+      event,
+      this.#currentResponseSnapshot,
+      this.#accumulatorContext,
+    );
     this.#currentResponseSnapshot = response;
     maybeEmit('event', event);
